+++ modules/pluggableGeometryExample/ospray/moduleInit.cpp
@@ -51,12 +51,14 @@ namespace ospray {
         (see comments regarding library name in CMakeLists.txt)
     */
 
-    extern "C" OSPError ospray_module_init_bilinear_patches(const int16_t &versionMajor,
-                                                            const int16_t &versionMinor)
+    extern "C" OSPError ospray_module_init_bilinear_patches(int16_t versionMajor,
+                                                            int16_t versionMinor,
+                                                            int16_t versionPatch)
     {
-      std::cout << "#osp: initializing the 'bilinear_patches' module" << std::endl;        
+      std::cout << "#osp: initializing the 'bilinear_patches' module" << std::endl;
       return moduleVersionCheck(versionMajor, versionMinor);
     }
+
   } // ::ospray::bilinearPatch
 } // ::ospray