+++ ospray/include/ospray/ospray.h
@@ -142,9 +142,6 @@ extern "C" {
   OSPRAY_INTERFACE OSPError ospInit(int *argc OSP_DEFAULT_VAL(NULL),
                                     const char **argv OSP_DEFAULT_VAL(NULL));
 
-  // return the OSPRay version of the installation
-  OSPRAY_INTERFACE int64_t ospVersionInfo();
-
   // returns the OSPRay Version in use by the device
   OSPRAY_INTERFACE int64_t ospDeviceGetProperty(OSPDevice, OSPDeviceProperty);