+++ CHANGELOG.md
@@ -67,9 +67,8 @@ Version History
 -   Removal of `cylinders` (grass) tutorial.
 -   `streamlines` of streamlines tutorial now connected
     `OSP_CATMULL_ROM` curves.
--   Changed signature for module init methods. New signature is:
-    `ospray_module_init_<module_name>(const int16_t&, const int16_t&)`
--   `moduleVersionCheck` method added to check module compatibility.
+-   Changed module init methods. New init methods have following signature:
+    `extern "C" OSPError ospray_module_init_<name>(int16_t versionMajor, int16_t versionMinor, int16_t versionPatch)`
 -   Known issues:
     -   Open VKL must be built with a compatible set of ISAs with OSPRay
         due to a known issue with VKL iterator types. This is easiest